refactor: change name to DSL::Style and extract more methods
